Revised modal popup functionality to be reusable. Combined functionality in SubmitIdeaButton.js and SubmitIdeasPopup.js into a single Popup.js file. Revised navbar.html, base.html, and idea_suggestion feature templates to implement this update. Added a smart-summary placeholder popup and corresponding button
