--- idea_suggestion/abstract_base_classes/__init__.py
@@ -1,1 +1,0 @@
-# __init__.py
