+++ templates/base.html
@@ -23,6 +23,6 @@
           {% block city_scripts %}{% endblock %}
         </main>
     </div>
-    <script src="{% static 'js/SubmitIdeasPopup.js' %}"></script>
+    <script src="{% static 'js/Popup.js' %}"></script>
 </body>
 </html>
