+++ templates/ideas/ideas_city_admin_inbox.html
@@ -31,9 +31,16 @@
         {% endfor %}
     </tbody>
 </table>
+
+
+
+
+
+
 {% endblock %}
 
 
 {% block city_scripts %}
     <script src="{% static 'js/IdeaInBoxLinks.js' %}"></script>
+    <script src="{% static 'js/SubmitIdeasPopup.js' %}"></script>
 {% endblock %}
