+++ templates/ideas/smart_summary.html
@@ -1,0 +1,18 @@
+
+{% load static %}
+
+<!-- Smart Summary Header -->
+<div class="modal-header">
+    <div class="title">Unread Inbox Summary</div>
+    <button data-close-x-button class="close-x-button">&times;</button>
+</div>
+
+<!-- Smart Summary Body -->
+<div class="modal-body">
+    Placeholder Text for Smart Summary
+</div>
+    
+<!-- Smart Summary Footer -->
+<div class="modal-footer">
+    <button data-close-button class="nav-btn" id="close-summary-modal" aria-label="Close">Close</button>
+</div>
